Refactor to interfaces
