+++ composer.json
@@ -23,8 +23,10 @@
         "php": "~5.5|7.*"
     },
     "require-dev": {
-        "laravel/framework": "5.*",
-        "phpunit/phpunit": "4.*",
+        "php": "^7.4",
+        "laravel/framework": "^8.0",
+        "orchestra/testbench": "^6.0",
+        "phpunit/phpunit": "^9.0",
         "friendsofphp/php-cs-fixer": "^2.15"
     },
     "autoload": {
@@ -34,7 +36,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Kyranb\\Footprints\\Test\\": "tests"
+            "Kyranb\\Footprints\\Tests\\": "tests/"
         }
     },
     "scripts": {
