+++ src/TrackingFilter.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace Kyranb\Footprints;
+
+use Illuminate\Http\Response;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
+class TrackingFilter implements TrackingFilterInterface
+{
+    /**
+     * The Request instance.
+     *
+     * @var \Illuminate\Http\Request
+     */
+    protected $request;
+
+    /**
+     * The Request instance.
+     *
+     * @var \Illuminate\Http\Response
+     */
+    protected $response;
+
+    /**
+     * Determine whether or not the request should be tracked.
+     *
+     * @param \Illuminate\Http\Request $request
+     * @param \Illuminate\Http\Response $response
+     * @return bool
+     */
+    public function shouldTrack(Request $request, Response $response)
+    {
+        $this->request = $request;
+        $this->response = $response;
+
+        //Only track get requests
+        if (! $this->request->isMethod('get')) {
+            return false;
+        }
+
+        if ($this->disableOnAuthentication()) {
+            return false;
+        }
+
+        if ($this->disableInternalLinks()) {
+            return false;
+        }
+
+        if ($this->disabledLandingPages($this->captureLandingPage())) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * @return bool
+     */
+    protected function disableOnAuthentication()
+    {
+        if (Auth::guard(config('footprints.guard'))->check() && config('footprints.disable_on_authentication')) {
+            return true;
+        }
+    }
+
+    /**
+     * @return bool
+     */
+    protected function disableInternalLinks()
+    {
+        if (! config('footprints.disable_internal_links')) {
+            return false;
+        }
+
+        $referrer_domain = parse_url($this->request->headers->get('referer'));
+        $referrer_domain = ! isset($referrer_domain['host']) ? null : $referrer_domain['host'];
+        $request_domain = $this->request->server('SERVER_NAME');
+
+        if (! empty($referrer_domain) && $referrer_domain == $request_domain) {
+            return true;
+        }
+    }
+
+    /**
+     *
+     * @param   string  $landing_page
+     * @return  array|boolean
+     */
+    protected function disabledLandingPages($landing_page = null)
+    {
+        $blacklist = (array) config('footprints.landing_page_blacklist');
+
+        if ($landing_page) {
+            $k = array_search($landing_page, $blacklist);
+
+            return $k === false ? false : true;
+        } else {
+            return $blacklist;
+        }
+    }
+
+    /**
+     * @return string
+     */
+    protected function captureLandingPage()
+    {
+        return $this->request->path();
+    }
+}
