--- src/tests/CaptureAttributionDataMiddlewareTest.php
@@ -1,3 +1,0 @@
-<?php
-
-//TODO ;)
