+++ tests/Unit/Middleware/CaptureAttributionDataMiddlewareTest.php
@@ -1,0 +1,55 @@
+<?php
+
+namespace Kyranb\Footprints\tests\Unit\Middleware;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Kyranb\Footprints\Middleware\CaptureAttributionDataMiddleware;
+use Kyranb\Footprints\TrackingFilterInterface;
+use Kyranb\Footprints\TrackingLoggerInterface;
+use PHPUnit\Framework\TestCase;
+
+class CaptureAttributionDataMiddlewareTest extends TestCase
+{
+    public function test_logs_when_filter_returns_true()
+    {
+        $request = new Request;
+        $response = new Response;
+
+        $trackingFilter = \Mockery::mock(TrackingFilterInterface::class);
+        $trackingFilter->shouldReceive('shouldTrack')
+            ->once()
+            ->andReturnTrue();
+
+        $trackingLogger = \Mockery::mock(TrackingLoggerInterface::class);
+        $trackingLogger->shouldReceive('track')
+            ->with($request, $response)
+            ->andReturn($response);
+
+        $middleware = new CaptureAttributionDataMiddleware($trackingFilter, $trackingLogger);
+
+        $this->assertEquals($response, $middleware->handle($request, function ($req) use ($response) {
+            return $response;
+        }));
+    }
+
+    public function test_do_not_log_when_filter_returns_false()
+    {
+        $request = new Request;
+        $response = new Response;
+
+        $trackingFilter = \Mockery::mock(TrackingFilterInterface::class);
+        $trackingFilter->shouldReceive('shouldTrack')
+            ->once()
+            ->andReturnFalse();
+
+        $trackingLogger = \Mockery::mock(TrackingLoggerInterface::class);
+        $trackingLogger->shouldNotReceive('track');
+
+        $middleware = new CaptureAttributionDataMiddleware($trackingFilter, $trackingLogger);
+
+        $this->assertEquals($response, $middleware->handle($request, function ($req) use ($response) {
+            return $response;
+        }));
+    }
+}
