+++ tests/Unit/TrackingFilterTest.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Kyranb\Footprints\tests\Unit;
+
+class TrackingFilterTest
+{
+    public function test_disabled_for_post_requests()
+    {
+        // TODO: Implement
+    }
+
+    public function test_disabled_on_authentication()
+    {
+        // TODO: Implement
+    }
+
+    public function test_disabled_for_internal_links()
+    {
+        // TODO: Implement
+    }
+
+    public function test_disabled_for_landing_page()
+    {
+        // TODO: Implement
+    }
+}
