+++ tests/Unit/TrackingLoggerTest.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Kyranb\Footprints\tests\Unit;
+
+class TrackingLoggerTest
+{
+    public function test_logging_job_handled_async()
+    {
+        // TODO: Test
+    }
+
+    public function test_logging_job_queued()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_ip()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_landing_domain()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_landing_page()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_landing_params()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_referrer()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_utm()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_referral()
+    {
+        // TODO: Test
+    }
+
+    public function test_attribution_data_custom()
+    {
+        // TODO: Test
+    }
+}
